+++ forge/src/main/java/net/mehvahdjukaar/supplementaries/common/events/forge/ServerEventsForge.java
@@ -252,7 +252,7 @@ public class ServerEventsForge {
             int t = slimed.supp$getSlimedTicks();
             if (t > 0) {
                 slimed.supp$setSlimedTicks(t - 1);
-            }else if(event.getEntity().level().random.nextFloat()<0.1){
+            }else if(event.getEntity().level().random.nextFloat()<0.001){
                 slimed.supp$setSlimedTicks(400);
             }
         }